Replace http_request with a kore_server name.
Instead of passing a http_request pointer, one
should now pass the name of a kore_server in order
to render partials that refer to files.
The kore_server must have tls enabled.
If NULL is passed or the kore_server has no tls enabled
the partials with files won't be rendered.